// demos/ui-extended/src/components/schema-builder/app/types.ts
export interface SectionPropsType {
  name: string;
  required: boolean;
  schema: { [key: string]: any };
  uischema: { [key: string]: any };
  onChange: (
    schema: { [key: string]: any },
    uischema: { [key: string]: any },
    ref?: string,
  ) => void;
  onNameChange: (arg0: string) => void;
  onDependentsChange: (
    arg0: {
      children: string[];
      value?: any;
    }[],
  ) => void;
  onRequireToggle: () => any;
  onDelete: () => any;
  onMoveUp?: () => any;
  onMoveDown?: () => any;
  path: string;
  definitionData: { [key: string]: any };
  definitionUi: { [key: string]: any };
  dependents?: Array<{
    children: Array<string>;
    value?: any;
  }>;
  parentProperties: AddFormObjectParametersType;
  neighborNames?: Array<string>;
  cardOpen: boolean;
  setCardOpen: (newState: boolean) => void;
  allFormInputs: { [key: string]: FormInput };
  categoryHash: { [key: string]: string };
  hideKey?: boolean;
  reference?: string;
  dependent?: boolean;
  parent?: string;
  mods?: Mods;
}
// ...
interface FormInputType {
  displayName: string;
  // given a data and ui schema, determine if the object is of this input type
  matchIf: Array<MatchType>;
  // allowed keys for ui:options
  possibleOptions?: Array<string>;
  defaultDataSchema: {
    [key: string]: any;
  };
  defaultUiSchema: {
    [key: string]: any;
  };
  // the data schema type
  type: DataType;
  // inputs on the preview card
  cardBody: CardComponentType;
  // inputs for the modal
  modalBody?: CardComponentType;
  [key: string]: any
}

export interface DataOptions {
  title: string;
  type?: string;
  description?: string;
  $ref?: string;
  default?: string | number;
}

export interface ModLabels {
  formNameLabel?: string;
  formDescriptionLabel?: string;
  objectNameLabel?: string;
  displayNameLabel?: string;
  descriptionLabel?: string;
  inputTypeLabel?: string;
  addElementLabel?: string;
  addSectionLabel?: string;
}

// optional properties that can add custom features to the form builder
export interface Mods {
  customFormInputs?: {
    [key: string]: FormInputType;
  };
  components?: {
    add?: (properties?: {
      [key: string]: any;
    }) => ReactElement | ReactElement[] | [];
  };
  tooltipDescriptions?: {
    add?: string;
    cardObjectName?: string;
    cardDisplayName?: string;
    cardDescription?: string;
    cardInputType?: string;
    cardSectionObjectName?: string;
    cardSectionDisplayName?: string;
    cardSectionDescription?: string;
  };
  labels?: ModLabels;
  showFormHead?: boolean;
  deactivatedFormInputs?: Array<string>;
  newElementDefaultDataOptions?: DataOptions;
  newElementDefaultUiSchema?: { [key: string]: any };
}

export type FormInput = FormInputType;
// ...
export interface AddFormObjectParametersType {
  schema: { [key: string]: any };
  uischema: { [key: string]: any };
  mods?: Mods;
  onChange: (
    schema: { [key: string]: any },
    uischema: { [key: string]: any },
  ) => any;
  definitionData: { [key: string]: any };
  definitionUi: { [key: string]: any };
  index?: number;
  categoryHash: { [key: string]: string };
}